Compile fix for linux

+++ renderdoc/driver/shaders/spirv/spirv_disassemble.cpp
@@ -139,7 +139,7 @@ rdcstr Reflector::Disassemble(const rdcstr &entryPoint,
       {
         rdcstr escaped = "\"";
 
-        for(int i = 0; i < ret.size(); i++)
+        for(size_t i = 0; i < ret.size(); i++)
         {
           if(i == 100 && i + 1 < ret.size())
           {
